Trying to suppress warning with a typecast.
git-svn-id: http://svn.seqan.de/seqan/trunk@7949 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/basic/basic_aggregates.h
@@ -367,7 +367,7 @@ namespace SEQAN_NAMESPACE_MAIN
         }
 		template <typename TPos>
         inline const _T& operator[](TPos k) const {
-            SEQAN_ASSERT(k >= 0 && k < size);
+            SEQAN_ASSERT(static_cast<__int64>(k) >= 0 && k < size);
             return i[k];
         }
 		inline _T* operator&() { return i; }
